Move auth properties and bearer token file interactions to separate files
